publisher/requirements.txt
# JIMBO77 Publisher - Python Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
aiohttp>=3.9.0
requests>=2.31.0
openai>=1.3.0
python-dotenv>=1.0.0
replicate>=0.25.0
